refactor: 렌더링하는 함수 파일로 분리 및 검색 기능의 공통 함수 분리
